FilesSyncHelperTest test done This tests the method InsertAllDBEntries from FilesSyncHelper, but ONLY for non-media Types (no video / audio folders). It creates a file tree on sd card, adds folders to database by method, and tests, if all needed directories are stored in database.

+++ app/src/androidTest/java/com/owncloud/android/utils/FilesSyncHelperTest.java
@@ -22,73 +22,46 @@
 
 package com.owncloud.android.utils;
 import android.content.ContentResolver;
-import android.content.Context;
 import android.database.Cursor;
-import android.util.Log;
-
 import com.nextcloud.client.account.MockUser;
-import com.nextcloud.client.account.User;
 import com.nextcloud.client.core.Clock;
-import com.nextcloud.client.core.ClockImpl;
 import com.nextcloud.client.preferences.AppPreferences;
-import com.nextcloud.client.preferences.AppPreferencesImpl;
 import com.owncloud.android.AbstractIT;
 import com.owncloud.android.MainApp;
-import com.owncloud.android.datamodel.FileSystemDataSet;
-import com.owncloud.android.datamodel.FilesystemDataProvider;
 import com.owncloud.android.datamodel.MediaFolderType;
 import com.owncloud.android.datamodel.SyncedFolder;
 import com.owncloud.android.datamodel.SyncedFolderProvider;
 import com.owncloud.android.db.ProviderMeta;
 import com.owncloud.android.files.services.FileUploader;
 import com.owncloud.android.files.services.NameCollisionPolicy;
-import com.owncloud.android.lib.common.utils.Log_OC;
-
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
-
 import static org.junit.Assert.fail;
 
 public class FilesSyncHelperTest extends AbstractIT{
-
     @Mock AppPreferences preferences;
     @Mock Clock clock;
 
-/*
-    @Mock
-    AppPreferences preferences;
-    @Mock
-    Clock clock;
-    @Mock FilesystemDataProvider provider;
-    @Mock ContentResolver resolver;
-    @Mock Context appContext;
-*/
     @Before
     public void init(){
         this.clock = Mockito.mock(Clock.class);
-        Mockito.when(clock.getCurrentTime()).thenReturn(1000L);
-        /*
-        appContext = Mockito.mock(Context.class);
-        resolver = Mockito.mock(ContentResolver.class);
-        provider = new FilesystemDataProvider(resolver);
-        clock = Mockito.mock(Clock.class);
-        Mockito.when(clock.getCurrentTime()).thenReturn(10000L);*/
+        Mockito.when(clock.getCurrentTime()).thenReturn(0L);
     }
-
+    @SuppressWarnings("ResultOfMethodCallIgnored") //file creation / deletion
     @Test
-    public void testInsertAllDBEntries(){
+    public void testInsertAllDBEntries(){ //Warning: this only tests for non-media folders
+        List<String> neededPaths = new LinkedList<>();
+
         String fileRoot = "/storage/emulated/0"; //replace this if test don't work
         try{
             File root = new File(fileRoot+"/nextcloudTest");
@@ -96,20 +69,30 @@ public class FilesSyncHelperTest extends AbstractIT{
             new File(fileRoot+"/nextcloudTest/images").mkdir();
             new File(fileRoot+"/nextcloudTest/videos").mkdir();
             new File(fileRoot+"/nextcloudTest/other").mkdir();
+            neededPaths.add(fileRoot+"/nextcloudTest/other");
             //add sample elements to file
             File image1 = new File(fileRoot+"/nextcloudTest/images/img1.png");
             image1.createNewFile();
-            new FileWriter(image1).write("Test");
             File video1 = new File(fileRoot+"/nextcloudTest/videos/vid1.mp4");
             video1.createNewFile();
-            new FileWriter(video1).write("Test");
             File other1 = new File(fileRoot+"/nextcloudTest/other/other1.txt");
-            new FileWriter(other1).write("Test");
+            other1.createNewFile();
+            neededPaths.add(fileRoot+"/nextcloudTest/other/other1.txt");
+            StringBuilder nestedpath = new StringBuilder(fileRoot + "/nextcloudTest/other");
+            for(int i = 0; i < 10; i++){
+                File add = new File(nestedpath+"/file"+i);
+                add.createNewFile();
+                neededPaths.add(add.getAbsolutePath());
+                File directory = new File(nestedpath+"/d"+i);
+                neededPaths.add(directory.getAbsolutePath());
+                directory.mkdir();
+                nestedpath.append("/d").append(i);
+            }
         }catch (IOException e){
             fail("Unable to create Test files: "+e);
         }
 
-        SyncedFolder images = new SyncedFolder(
+        /*SyncedFolder images = new SyncedFolder(
             fileRoot+"/nextcloudTest/images",
             "/testing/images",
             false,
@@ -138,7 +121,7 @@ public class FilesSyncHelperTest extends AbstractIT{
             0L,
             MediaFolderType.VIDEO,
             false
-        );
+        );*/
         SyncedFolder other = new SyncedFolder(
             fileRoot+"/nextcloudTest/other",
             "/testing/images",
@@ -155,20 +138,55 @@ public class FilesSyncHelperTest extends AbstractIT{
             false
         );
         SyncedFolderProvider provider = new SyncedFolderProvider(MainApp.getAppContext().getContentResolver(),preferences,clock);
-        provider.storeSyncedFolder(images);
-        provider.storeSyncedFolder(videos);
+        //provider.storeSyncedFolder(images);
+        //provider.storeSyncedFolder(videos);
         provider.storeSyncedFolder(other);
 
-        //Set Up mock query
-
+        //Execute Test
         FilesSyncHelper.insertAllDBEntries(preferences,clock,false);
-        FilesystemDataProvider fileprovider = new FilesystemDataProvider(MainApp.getAppContext().getContentResolver());
-        Log.d("FilesSyncHelperTest",fileprovider.getFilesForUpload("%",images.getId()+"").size()+"sized");
 
+        //got from filesystemdataprovider to get all synced folders
+        Set<String> filesInDB = new HashSet<>();
+        ContentResolver resolver = MainApp.getAppContext().getContentResolver();
+        Cursor cursor = resolver.query(
+            ProviderMeta.ProviderTableMeta.CONTENT_URI_FILESYSTEM,
+            null,
+            null,
+            null,
+            null);
+        if (cursor != null) {
+            if (cursor.moveToFirst()) {
+                do {
+                    String value = cursor.getString(cursor.getColumnIndexOrThrow(
+                        ProviderMeta.ProviderTableMeta.FILESYSTEM_FILE_LOCAL_PATH));
+                    if (value == null) {
+                        fail("Cannot get local path");
+                    } else {
+                        File file = new File(value);
+                        if (!file.exists()) {
+                            fail("Ignoring file for upload (doesn't exist): " + value);
+                        } else if (!SyncedFolderUtils.isQualifiedFolder(file.getParent())) {
+                            fail("Ignoring file for upload (unqualified folder): " + value);
+                        } else if (!SyncedFolderUtils.isFileNameQualifiedForAutoUpload(file.getName())) {
+                            fail("Ignoring file for upload (unqualified file): " + value);
+                        } else {
+                            filesInDB.add(value);
+                        }
+                    }
+                } while (cursor.moveToNext());
+            }
+            cursor.close();
+        }
+        if(!filesInDB.containsAll(neededPaths)){
+            fail("Not all files from synced folder are correctly in database");
+        }
 
         //cleanup (I don't know if it is done automatically, so I do it myself)
         provider.deleteSyncFoldersForAccount(new MockUser("FilesSyncHelperTest",MockUser.DEFAULT_MOCK_ACCOUNT_TYPE));
-        boolean deleted = true;
+        Collections.reverse(neededPaths); //delete in reverse
+        for(String path: neededPaths){
+            new File(path).delete();
+        }
         new File(fileRoot+"/nextcloudTest/images/img1.png").delete();
         new File(fileRoot+"/nextcloudTest/videos/vid1.mp4").delete();
         new File(fileRoot+"/nextcloudTest/other/other1.txt").delete();
@@ -176,6 +194,5 @@ public class FilesSyncHelperTest extends AbstractIT{
         new File(fileRoot+"/nextcloudTest/videos").delete();
         new File(fileRoot+"/nextcloudTest/other").delete();
         new File(fileRoot+"/nextcloudTest").delete();
-
     }
 }
